Refactored `eggScnRenderer`

+++ include/egg/gfxe/eggDrawPathBase.h
@@ -13,7 +13,7 @@ class DrawPathBase : public ScreenEffectBase, public IScnProc {
 public:
     DrawPathBase();
 
-    virtual int getNumStep() const = 0;      // at 0x1C
+    virtual u16 getNumStep() const = 0;      // at 0x1C
     virtual void internalResetForDraw() {}   // at 0x20
     virtual void internalCalc() {}           // at 0x24
     virtual void internalDraw(u16 step) = 0; // at 0x28
